refactor: models are now referred to as "data" and contains more properties This is an attempt to improve compatibility by consolidating chat messages and custom roll. The field parsing and data entry construction is now exposed by BetterRolls.builder.constructX(). Referring to the models as "data" makes the api somewhat similar to how Foundry entities work. Eventually a wrapper object over this data will be constructed to facilitate interaction with it.
